+++ jest.config.ts
@@ -76,86 +76,21 @@ export default {
   // maxWorkers: '50%',
 
   // An array of directory names to be searched recursively up from the requiring module's location
-  // moduleDirectories: [
-  //   'node_modules'
-  // ],
-
-  // An array of file extensions your modules use
-  // moduleFileExtensions: [
-  //   'js',
-  //   'mjs',
-  //   'cjs',
-  //   'jsx',
-  //   'ts',
-  //   'tsx',
-  //   'json',
-  //   'node'
-  // ],
+  moduleDirectories: [
+    'node_modules',
+    '<rootDir>/src'
+  ],
 
   // A map from regular expressions to module names or to arrays of module names that allow to stub out resources with a single module
-  // moduleNameMapper: {},
-
-  // An array of regexp pattern strings, matched against all module paths before considered 'visible' to the module loader
-  // modulePathIgnorePatterns: [],
-
-  // Activates notifications for test results
-  // notify: false,
-
-  // An enum that specifies notification mode. Requires { notify: true }
-  // notifyMode: 'failure-change',
+  moduleNameMapper: {
+    "^@archwayhq/arch3-proto/(.*)$": "<rootDir>/../arch3-proto/$1"
+  },
 
   // A preset that is used as a base for Jest's configuration
   preset: 'ts-jest',
 
-  // Run tests from one or more projects
-  // projects: undefined
-
-  // Use this configuration option to add custom reporters to Jest
-  // reporters: undefined,
-
-  // Automatically reset mock state before every test
-  // resetMocks: false,
-
-  // Reset the module registry before running each individual test
-  // resetModules: false,
-
-  // A path to a custom resolver
-  // resolver: undefined,
-
-  // Automatically restore mock state and implementation before every test
-  // restoreMocks: false,
-
-  // The root directory that Jest should scan for tests and modules within
-  // rootDir: undefined,
-
-  // A list of paths to directories that Jest should use to search for files in
-  // roots: [
-  //   '<rootDir>'
-  // ],
-
-  // Allows you to use a custom runner instead of Jest's default test runner
-  // runner: 'jest-runner',
-
-  // The paths to modules that run some code to configure or set up the testing environment before each test
-  // setupFiles: [],
-
-  // A list of paths to modules that run some code to configure or set up the testing framework before each test
-  // setupFilesAfterEnv: [],
-
-  // The number of seconds after which a test is considered as slow and reported as such in the results.
-  // slowTestThreshold: 5,
-
-  // A list of paths to snapshot serializer modules Jest should use for snapshot testing
-  // snapshotSerializers: [],
-
   // The test environment that will be used for testing
   testEnvironment: 'node',
-
-  // Options that will be passed to the testEnvironment
-  // testEnvironmentOptions: {},
-
-  // Adds a location field to test results
-  // testLocationInResults: false,
 
   // The glob patterns Jest uses to detect test files
   testMatch: [
@@ -169,25 +104,12 @@ export default {
     '.history',
   ],
 
-  // The regexp pattern or array of patterns that Jest uses to detect test files
-  // testRegex: [],
-
-  // This option allows the use of a custom results processor
-  // testResultsProcessor: undefined,
-
-  // This option allows use of a custom test runner
-  // testRunner: 'jest-circus/runner',
-
   // Default timeout of a test in milliseconds.
   testTimeout: 3000,
 
-  // A map from regular expressions to paths to transformers
-  // transform: undefined,
-
   // An array of regexp pattern strings that are matched against all source file paths, matched files will skip transformation
   // transformIgnorePatterns: [
-  //   '/node_modules/',
-  //   '\\.pnp\\.[^\\/]+$'
+  //   "<rootDir>/src/(?!archwayclient.ts/)"
   // ],
 
   // An array of regexp pattern strings that are matched against all modules before the module loader will automatically return a mock for them